Rename nav.js to filterItems.js
